+++ src/functions/utils/assembleSupSub.js
@@ -1,6 +1,7 @@
 // @flow
 import buildCommon from "../../buildCommon";
 import * as html from "../../buildHTML";
+import utils from "../../utils";
 import type {StyleInterface} from "../../Style";
 import type Options from "../../Options";
 import type {DomSpan, SymbolNode} from "../../domTree";
@@ -18,6 +19,7 @@ export const assembleSupSub = (
     baseShift: number,
 ): DomSpan => {
     base = buildCommon.makeSpan([], [base]);
+    const subIsSingleCharacter = subGroup &&  utils.isCharacterBox(subGroup);
     let sub;
     let sup;
     // We manually have to handle the superscripts and subscripts. This,
@@ -105,6 +107,13 @@ export const assembleSupSub = (
         return base;
     }
 
-    return buildCommon.makeSpan(
-        ["mop", "op-limits"], [finalGroup], options);
+    const parts = [finalGroup];
+    if (sub && slant !== 0 && !subIsSingleCharacter) {
+        // A negative margin-left was applied to the lower limit.
+        // Avoid an overlap by placing a spacer on the left on the group.
+        const spacer = buildCommon.makeSpan(["mspace"], [], options);
+        spacer.style.marginRight = `${slant}em`;
+        parts.unshift(spacer);
+    }
+    return buildCommon.makeSpan(["mop", "op-limits"], parts, options);
 };